Refactor upload_raw_package from pip to general
